requirements.txt
fastapi 
uvicorn 
psycopg2-binary 
sqlalchemy 
pydantic
pydantic-settings
python-dotenv
alembic
requests
passlib[bcrypt]
python-jose[cryptography]
python-multipart
slowapi
redis
streamlit>=1.28.0
requests>=2.31.0
folium>=0.14.0
streamlit-folium>=0.15.0
pandas>=2.0.0